% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcn.R
\name{calcn}
\alias{calcn}
\title{Calculates n Given A Three Element Vector}
\usage{
calcn(x, n)
}
\arguments{
\item{x}{\itemize{
\item a 3 element vector
}}

\item{n}{\itemize{
\item an integer
}}
}
\value{
\itemize{
\item a numeric value
}
}
\description{
Calculates n Given A Three Element Vector
}
\examples{
library(hw04adim7calcnpltr)
x = c(2, 4, 3), n = 3
Calculate sequence
calcn(c(2,3,3), 3L) # Run
}
