% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcn_lineplot.R
\name{calcn_lineplot}
\alias{calcn_lineplot}
\title{Line Plot from recursive function
Takes output of the calcn() applied o a tibble and creates a line plot}
\usage{
calcn_lineplot(df)
}
\arguments{
\item{df}{a n x 4 tibble}
}
\value{
a line plot
}
\description{
Line Plot from recursive function
Takes output of the calcn() applied o a tibble and creates a line plot
}
