% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04adim7calcnpltr-package.R
\docType{package}
\name{hw04adim7calcnpltr-package}
\alias{hw04adim7calcnpltr}
\alias{hw04adim7calcnpltr-package}
\title{hw04adim7calcnpltr: Calculates n Given A Three Element Vector or Tibble and Creates a Line Plot}
\description{
Takes a 4 column tibble and calculates n as output, then creates a line plot.
}
\author{
\strong{Maintainer}: Adim Okwudishu \email{adimkris@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
